separated calibration types
