+++ camera_calibration/src/camera_calibration/camera_calibrator.py
@@ -40,7 +40,9 @@ import sensor_msgs.msg
 import sensor_msgs.srv
 import threading
 import time
-from camera_calibration.calibrator import MonoCalibrator, StereoCalibrator, Patterns
+from camera_calibration.mono_calibrator import MonoCalibrator
+from camera_calibration.stereo_calibrator import StereoCalibrator
+from camera_calibration.calibrator import Patterns
 try:
     from queue import Queue
 except ImportError:
@@ -54,6 +56,7 @@ class BufferQueue(Queue):
     """Slight modification of the standard Queue that discards the oldest item
     when adding an item and the queue is full.
     """
+
     def put(self, item, *args, **kwargs):
         # The base implementation, for reference:
         # https://github.com/python/cpython/blob/2.7/Lib/Queue.py#L107
@@ -92,14 +95,17 @@ class ConsumerThread(threading.Thread):
 
 
 class CalibrationNode(Node):
-    def __init__(self, name, boards, service_check = True, synchronizer = message_filters.TimeSynchronizer, flags = 0,
-                 fisheye_flags = 0, pattern=Patterns.Chessboard, camera_name='', checkerboard_flags = 0,
-                 max_chessboard_speed = -1, queue_size = 1):
+    def __init__(self, name, boards, service_check=True, synchronizer=message_filters.TimeSynchronizer, flags=0,
+                 fisheye_flags=0, pattern=Patterns.Chessboard, camera_name='', checkerboard_flags=0,
+                 max_chessboard_speed=-1, queue_size=1):
         super().__init__(name)
 
-        self.set_camera_info_service = self.create_client(sensor_msgs.srv.SetCameraInfo, "camera/set_camera_info")
-        self.set_left_camera_info_service = self.create_client(sensor_msgs.srv.SetCameraInfo, "left_camera/set_camera_info")
-        self.set_right_camera_info_service = self.create_client(sensor_msgs.srv.SetCameraInfo, "right_camera/set_camera_info")
+        self.set_camera_info_service = self.create_client(
+            sensor_msgs.srv.SetCameraInfo, "camera/set_camera_info")
+        self.set_left_camera_info_service = self.create_client(
+            sensor_msgs.srv.SetCameraInfo, "left_camera/set_camera_info")
+        self.set_right_camera_info_service = self.create_client(
+            sensor_msgs.srv.SetCameraInfo, "right_camera/set_camera_info")
 
         if service_check:
             available = False
@@ -123,12 +129,15 @@ class CalibrationNode(Node):
         self._pattern = pattern
         self._camera_name = camera_name
         self._max_chessboard_speed = max_chessboard_speed
-        lsub = message_filters.Subscriber(self, sensor_msgs.msg.Image, 'left', qos_profile=self.get_topic_qos("left"))
-        rsub = message_filters.Subscriber(self, sensor_msgs.msg.Image, 'right', qos_profile=self.get_topic_qos("right"))
+        lsub = message_filters.Subscriber(
+            self, sensor_msgs.msg.Image, 'left', qos_profile=self.get_topic_qos("left"))
+        rsub = message_filters.Subscriber(
+            self, sensor_msgs.msg.Image, 'right', qos_profile=self.get_topic_qos("right"))
         ts = synchronizer([lsub, rsub], 4)
         ts.registerCallback(self.queue_stereo)
 
-        msub = message_filters.Subscriber(self, sensor_msgs.msg.Image, 'image', qos_profile=self.get_topic_qos("image"))
+        msub = message_filters.Subscriber(
+            self, sensor_msgs.msg.Image, 'image', qos_profile=self.get_topic_qos("image"))
         msub.registerCallback(self.queue_monocular)
 
         self.q_mono = BufferQueue(queue_size)
@@ -148,6 +157,7 @@ class CalibrationNode(Node):
 
     def redraw_stereo(self, *args):
         pass
+
     def redraw_monocular(self, *args):
         pass
 
@@ -162,11 +172,11 @@ class CalibrationNode(Node):
             if self._camera_name:
                 self.c = MonoCalibrator(self._boards, self._calib_flags, self._fisheye_calib_flags, self._pattern, name=self._camera_name,
                                         checkerboard_flags=self._checkerboard_flags,
-                                        max_chessboard_speed = self._max_chessboard_speed)
+                                        max_chessboard_speed=self._max_chessboard_speed)
             else:
                 self.c = MonoCalibrator(self._boards, self._calib_flags, self._fisheye_calib_flags, self._pattern,
                                         checkerboard_flags=self.checkerboard_flags,
-                                        max_chessboard_speed = self._max_chessboard_speed)
+                                        max_chessboard_speed=self._max_chessboard_speed)
 
         # This should just call the MonoCalibrator
         drawable = self.c.handle_msg(msg)
@@ -178,16 +188,15 @@ class CalibrationNode(Node):
             if self._camera_name:
                 self.c = StereoCalibrator(self._boards, self._calib_flags, self._fisheye_calib_flags, self._pattern, name=self._camera_name,
                                           checkerboard_flags=self._checkerboard_flags,
-                                          max_chessboard_speed = self._max_chessboard_speed)
+                                          max_chessboard_speed=self._max_chessboard_speed)
             else:
                 self.c = StereoCalibrator(self._boards, self._calib_flags, self._fisheye_calib_flags, self._pattern,
                                           checkerboard_flags=self._checkerboard_flags,
-                                          max_chessboard_speed = self._max_chessboard_speed)
+                                          max_chessboard_speed=self._max_chessboard_speed)
 
         drawable = self.c.handle_msg(msg)
         self.displaywidth = drawable.lscrib.shape[1] + drawable.rscrib.shape[1]
         self.redraw_stereo(drawable)
-
 
     def check_set_camera_info(self, response):
         if response.success:
@@ -196,12 +205,14 @@ class CalibrationNode(Node):
         for i in range(10):
             print("!" * 80)
         print()
-        print("Attempt to set camera info failed: " + response.result() if response.result() is not None else "Not available")
+        print("Attempt to set camera info failed: " + response.result()
+              if response.result() is not None else "Not available")
         print()
         for i in range(10):
             print("!" * 80)
         print()
-        self.get_logger().error('Unable to set camera info for calibration. Failure message: %s' % response.result() if response.result() is not None else "Not available")
+        self.get_logger().error('Unable to set camera info for calibration. Failure message: %s' %
+                                response.result() if response.result() is not None else "Not available")
         return False
 
     def do_upload(self):
@@ -240,7 +251,8 @@ class CalibrationNode(Node):
             qos_profile.depth = qos_profile_system_default.depth
             return qos_profile
         else:
-            self.get_logger().warn(f"No publishers available for topic {topic_name}. Using system default QoS for subscriber.")
+            self.get_logger().warn(
+                f"No publishers available for topic {topic_name}. Using system default QoS for subscriber.")
             return qos_profile_system_default
 
 
@@ -276,12 +288,14 @@ class OpenCVCalibrationNode(CalibrationNode):
     def initWindow(self):
         cv2.namedWindow("display", cv2.WINDOW_NORMAL)
         cv2.setMouseCallback("display", self.on_mouse)
-        cv2.createTrackbar("Camera type: \n 0 : pinhole \n 1 : fisheye", "display", 0,1, self.on_model_change)
+        cv2.createTrackbar("Camera type: \n 0 : pinhole \n 1 : fisheye",
+                           "display", 0, 1, self.on_model_change)
         cv2.createTrackbar("scale", "display", 0, 100, self.on_scale)
 
     @classmethod
-    def putText(cls, img, text, org, color = (0,0,0)):
-        cv2.putText(img, text, org, cls.FONT_FACE, cls.FONT_SCALE, color, thickness = cls.FONT_THICKNESS)
+    def putText(cls, img, text, org, color=(0, 0, 0)):
+        cv2.putText(img, text, org, cls.FONT_FACE, cls.FONT_SCALE,
+                    color, thickness=cls.FONT_THICKNESS)
 
     @classmethod
     def getTextSize(cls, text):
@@ -293,7 +307,8 @@ class OpenCVCalibrationNode(CalibrationNode):
                 if 180 <= y < 280:
                     print("**** Calibrating ****")
                     # Perform calibration in another thread to prevent UI blocking
-                    threading.Thread(target=self.c.do_calibration, name="Calibration").start()
+                    threading.Thread(target=self.c.do_calibration,
+                                     name="Calibration").start()
                     self.buttons(self._last_display)
                     self.queue_display.put(self._last_display)
             if self.c.calibrated:
@@ -303,12 +318,14 @@ class OpenCVCalibrationNode(CalibrationNode):
                     # Only shut down if we set camera info correctly, #3993
                     if self.do_upload():
                         rclpy.shutdown()
+
     def on_model_change(self, model_select_val):
         if self.c == None:
             print("Cannot change camera model until the first image has been received")
             return
 
-        self.c.set_cammodel( CAMERA_MODEL.PINHOLE if model_select_val < 0.5 else CAMERA_MODEL.FISHEYE)
+        self.c.set_cammodel(
+            CAMERA_MODEL.PINHOLE if model_select_val < 0.5 else CAMERA_MODEL.FISHEYE)
 
     def on_scale(self, scalevalue):
         if self.c.calibrated:
@@ -321,15 +338,17 @@ class OpenCVCalibrationNode(CalibrationNode):
             color = (155, 155, 80)
         else:
             color = (224, 224, 224)
-        cv2.circle(dst, (size[0] // 2, size[1] // 2), min(size) // 2, color, -1)
+        cv2.circle(dst, (size[0] // 2, size[1] // 2),
+                   min(size) // 2, color, -1)
         (w, h) = self.getTextSize(label)
-        self.putText(dst, label, ((size[0] - w) // 2, (size[1] + h) // 2), (255,255,255))
+        self.putText(
+            dst, label, ((size[0] - w) // 2, (size[1] + h) // 2), (255, 255, 255))
 
     def buttons(self, display):
         x = self.displaywidth
-        self.button(display[180:280,x:x+100], "CALIBRATE", self.c.goodenough)
-        self.button(display[280:380,x:x+100], "SAVE", self.c.calibrated)
-        self.button(display[380:480,x:x+100], "COMMIT", self.c.calibrated)
+        self.button(display[180:280, x:x+100], "CALIBRATE", self.c.goodenough)
+        self.button(display[280:380, x:x+100], "SAVE", self.c.calibrated)
+        self.button(display[380:480, x:x+100], "COMMIT", self.c.calibrated)
 
     def y(self, i):
         """Set up right-size images"""
@@ -346,23 +365,25 @@ class OpenCVCalibrationNode(CalibrationNode):
         height = drawable.scrib.shape[0]
         width = drawable.scrib.shape[1]
 
-        display = numpy.zeros((max(480, height), width + 100, 3), dtype=numpy.uint8)
-        display[0:height, 0:width,:] = drawable.scrib
-        display[0:height, width:width+100,:].fill(255)
+        display = numpy.zeros(
+            (max(480, height), width + 100, 3), dtype=numpy.uint8)
+        display[0:height, 0:width, :] = drawable.scrib
+        display[0:height, width:width+100, :].fill(255)
 
         self.buttons(display)
         if not self.c.calibrated:
             if drawable.params:
-                 for i, (label, lo, hi, progress) in enumerate(drawable.params):
-                    (w,_) = self.getTextSize(label)
-                    self.putText(display, label, (width + (100 - w) // 2, self.y(i)))
-                    color = (0,255,0)
+                for i, (label, lo, hi, progress) in enumerate(drawable.params):
+                    (w, _) = self.getTextSize(label)
+                    self.putText(display, label,
+                                 (width + (100 - w) // 2, self.y(i)))
+                    color = (0, 255, 0)
                     if progress < 1.0:
                         color = (0, int(progress*255.), 255)
                     cv2.line(display,
-                            (int(width + lo * 100), self.y(i) + 20),
-                            (int(width + hi * 100), self.y(i) + 20),
-                            color, 4)
+                             (int(width + lo * 100), self.y(i) + 20),
+                             (int(width + hi * 100), self.y(i) + 20),
+                             color, 4)
 
         else:
             self.putText(display, "lin.", (width, self.y(0)))
@@ -371,7 +392,7 @@ class OpenCVCalibrationNode(CalibrationNode):
                 msg = "?"
             else:
                 msg = "%.2f" % linerror
-                #print "linear", linerror
+                # print "linear", linerror
             self.putText(display, msg, (width, self.y(1)))
 
         self._last_display = display
@@ -381,25 +402,27 @@ class OpenCVCalibrationNode(CalibrationNode):
         height = drawable.lscrib.shape[0]
         width = drawable.lscrib.shape[1]
 
-        display = numpy.zeros((max(480, height), 2 * width + 100, 3), dtype=numpy.uint8)
-        display[0:height, 0:width,:] = drawable.lscrib
-        display[0:height, width:2*width,:] = drawable.rscrib
-        display[0:height, 2*width:2*width+100,:].fill(255)
+        display = numpy.zeros(
+            (max(480, height), 2 * width + 100, 3), dtype=numpy.uint8)
+        display[0:height, 0:width, :] = drawable.lscrib
+        display[0:height, width:2*width, :] = drawable.rscrib
+        display[0:height, 2*width:2*width+100, :].fill(255)
 
         self.buttons(display)
 
         if not self.c.calibrated:
             if drawable.params:
                 for i, (label, lo, hi, progress) in enumerate(drawable.params):
-                    (w,_) = self.getTextSize(label)
-                    self.putText(display, label, (2 * width + (100 - w) // 2, self.y(i)))
-                    color = (0,255,0)
+                    (w, _) = self.getTextSize(label)
+                    self.putText(display, label, (2 * width +
+                                 (100 - w) // 2, self.y(i)))
+                    color = (0, 255, 0)
                     if progress < 1.0:
                         color = (0, int(progress*255.), 255)
                     cv2.line(display,
-                            (int(2 * width + lo * 100), self.y(i) + 20),
-                            (int(2 * width + hi * 100), self.y(i) + 20),
-                            color, 4)
+                             (int(2 * width + lo * 100), self.y(i) + 20),
+                             (int(2 * width + hi * 100), self.y(i) + 20),
+                             color, 4)
 
         else:
             self.putText(display, "epi.", (2 * width, self.y(0)))
@@ -411,7 +434,8 @@ class OpenCVCalibrationNode(CalibrationNode):
             # TODO dim is never set anywhere. Supposed to be observed chessboard size?
             if drawable.dim != -1:
                 self.putText(display, "dim", (2 * width, self.y(2)))
-                self.putText(display, "%.3f" % drawable.dim, (2 * width, self.y(3)))
+                self.putText(display, "%.3f" %
+                             drawable.dim, (2 * width, self.y(3)))
 
         self._last_display = display
         self.queue_display.put(display)
